TableGen: Handle nontrivial foreach range bounds
This allows using anything that isn't a literal integer as the bounds
for a foreach. Some of the diagnostics aren't perfect, but nobody ever
accused tablegen of having good errors. For example, the existing
wording suggests a bitrange is valid, but as far as I can tell this
has never worked.

Fixes bug 41958.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@361434 91177308-0d34-0410-b5e6-96231b3b80d8