[X86] Add NEG to isUseDefConvertible.
We can use the C flag from NEG to detect that the input was zero.

Really we could probably use the Z flag too. But C matches what
we'd do for usubo 0, X.

Haven't found a test case for this due to the usubo formation
in CGP. But I verified if I comment out the CGP code this
transformation catches some of the same cases.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@365929 91177308-0d34-0410-b5e6-96231b3b80d8