[RISCV] Attempt to make rv{32,64}i-aliases-invalid.s less flaky
These tests have been disabled on Linux and Windows due to failing
there. I think that could be down to a race condition between stdout
and stderr, so I have disabled output to stdout.

For the moment, only re-enable on linux, because I don't have a windows
machine to test on.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@367311 91177308-0d34-0410-b5e6-96231b3b80d8