[llvm-dlltool] Clarify an error message. NFC.
The parameter to the -D (--dllname) option is the name of the dll
that llvm-dlltool produces an import library for. Even though this
is named "OutputFile" in the COFFModuleDefinition class, it's not
an output file name in the context of llvm-dlltool, but the name
of the DLL to create an import library for.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@367676 91177308-0d34-0410-b5e6-96231b3b80d8