NATIVE tablegen needs to depend on target tablegen
This dependency was removed in  r357486, which has lead to a stream of difficult to diagnose bugs.

Without this dependency, when building with `LLVM_OPTIMIZED_TABLEGEN=On` the native tablegen executible may not be rebuilt at all, and often won't get rebuilt before targets that use the tablegen headers. In the best case this results in a build-time failure, in the worst case it results in runtime failures.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@367895 91177308-0d34-0410-b5e6-96231b3b80d8