Support swifterror in coroutine lowering.
The support for swifterror allocas should work in all lowerings.
The support for swifterror arguments only really works in a lowering
with prototypes where you can ensure that the prototype also has a
swifterror argument; I'm not really sure how it could possibly be
made to work in the switch lowering.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@368795 91177308-0d34-0410-b5e6-96231b3b80d8