AMDGPU: Fix emitting multiple stack loads for stack passed workitems
The same stack is loaded for each workitem ID, and each use. Nothing
prevents you from creating multiple fixed stack objects with the same
offsets, so this was creating a load for each unique frame index,
despite them being the same offset. Re-use the same frame index so the
loads are CSEable.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@371148 91177308-0d34-0410-b5e6-96231b3b80d8